requirements.txt
Django==1.8.5
django-flat-theme==1.1.1
django-taggit==0.17.1
django-tinymce==2.0.5
django-wysiwyg==0.7.1
